fix error Validasi.php

+++ application/controllers/Adminlembaga.php
@@ -89,7 +89,9 @@ class Adminlembaga extends MY_Controller {
 
 			if ($kirim) {
 				
-				$this->db->where_in('kon_id',$kon_ids)->delete('m_admin');
+				if(count($kon_ids) > 0) {
+					$this->db->where_in('kon_id',$kon_ids)->delete('m_admin');	
+				}
 				$data_admin = [
 					'user_id' => $post['username'],
 					'username' => $post['email'],
