+++ application/controllers/Trainer.php
@@ -197,7 +197,10 @@ class Trainer extends MY_Controller {
 					}
 				}
 
-				$this->db->where_in('kon_id',$kon_ids)->delete('m_admin');
+				if(count($kon_ids) > 0) {
+					$this->db->where_in('kon_id',$kon_ids)->delete('m_admin');	
+				}
+				
 
                 $data_admin = [
                     'user_id'  => $post['username'],
